Fix #37676 - Throw  E_WARNING for invalid array container access
Commit will raise an E_WARNING when a user uses a scalar value
for read-access as an array.  This mimics what happens when you
use a scalar value as an array for write-access.  This does not
change how list() access arrays with regard to the E_WARNING.
RFC: https://wiki.php.net/rfc/notice-for-non-valid-array-container