Implement Comparison Chaining
This feature implements support for arbitrary comparison chaining.  This
change sets the precedence for equality operations as well as comparison
operations to the same level.  This enables us to, parse in a
left-to-right fashion continuing to compare against the previous
operations right operator.  The implementation enables short-circuiting
in the event if a left operation is false, we will skip over further
right operations.